feat: Update meta tags and project data
Switches the image CDN for social media meta tags and favicons to postimg.cc for better performance.
Also, refactors the `ProjectItem` interface in `types.ts` to remove redundant fields (`title`, `category`, `location`), as this data is no longer used in the `constants.ts` file. The `PROJECTS` array in `constants.ts` is updated to reflect these changes, now only containing `id` and `image`.
Adjusts navbar and footer scroll offsets for better alignment with the fixed header. Enhances scroll animations on the Home and About pages for a smoother user experience.